requirements.txt
click==8.1.2
cycler==0.11.0
fonttools==4.32.0
imageio==2.16.1
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
matplotlib==3.5.1
networkx==2.8
numba==0.55.1
numpy==1.21.5
packaging==21.3
Pillow==9.1.0
pydantic==1.9.0
pyparsing==3.0.8
python-dateutil==2.8.2
PyWavelets==1.3.0
pyxelate==2.1.1
PyYAML==6.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.4.8
tqdm==4.64.0
typing_extensions==4.2.0
yml==0.0.1
